requirements.txt
asgiref==3.8.1
Django==5.0.4
django-admin-honeypot-updated-2021==1.2.0
django-ckeditor-5==0.2.12
django-filter==24.2
django-ipware==7.0.1
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-leaflet==0.29.1
djangorestframework==3.15.1
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg==3.1.18
PyJWT==2.8.0
python-ipware==3.0.0
sqlparse==0.4.4
typing_extensions==4.10.0
